[Refactor] Move connector sink composition to dedicated builders
