+++ be/src/exec/data_sinks/hive_sink_builder.h
@@ -1,0 +1,37 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#pragma once
+
+#include <vector>
+
+#include "common/status.h"
+#include "exec_primitive/pipeline/pipeline_fwd.h"
+#include "gen_cpp/DataSinks_types.h"
+#include "gen_cpp/Exprs_types.h"
+
+namespace starrocks {
+
+class HiveSinkBuilder {
+public:
+    explicit HiveSinkBuilder(const std::vector<TExpr>& output_exprs) : _output_exprs(output_exprs) {}
+
+    Status build(pipeline::OpFactories prev_operators, const TDataSink& thrift_sink,
+                 pipeline::PipelineBuilderContext* context) const;
+
+private:
+    const std::vector<TExpr>& _output_exprs;
+};
+
+} // namespace starrocks
