+++ be/src/exec/data_sinks/hive_table_sink.cpp
@@ -14,17 +14,13 @@
 
 #include "exec/data_sinks/hive_table_sink.h"
 
+#include <sstream>
+#include <utility>
+
 #include "common/runtime_profile.h"
-#include "connector/connector_registry.h"
-#include "exec/pipeline/fragment_context.h"
-#include "exec/pipeline/pipeline_builder.h"
-#include "exec/pipeline/pipeline_builder_operators.h"
-#include "exprs/expr.h"
+#include "exec/data_sinks/hive_sink_builder.h"
 #include "exprs/expr_executor.h"
-#include "exprs/expr_factory.h"
-#include "formats/csv/csv_defaults.h"
 #include "runtime/runtime_state.h"
-#include "runtime/service_contexts.h"
 
 namespace starrocks {
 
@@ -65,77 +61,7 @@ Status HiveTableSink::close(RuntimeState* state, const Status& exec_status) {
 
 Status HiveTableSink::decompose_to_pipeline(pipeline::OpFactories prev_operators, const TDataSink& thrift_sink,
                                             pipeline::PipelineBuilderContext* context) const {
-    auto runtime_state = context->runtime_state();
-    auto fragment_ctx = context->fragment_context();
-    auto output_exprs = this->get_output_expr();
-    const auto& t_hive_sink = thrift_sink.hive_table_sink;
-    const auto num_data_columns = t_hive_sink.data_column_names.size();
-    std::vector<TExpr> data_exprs(output_exprs.begin(), output_exprs.begin() + num_data_columns);
-    std::vector<TExpr> partition_exprs(output_exprs.begin() + num_data_columns, output_exprs.end());
-
-    auto sink_ctx = std::make_shared<connector::HiveChunkSinkContext>();
-    sink_ctx->path = t_hive_sink.staging_dir;
-    sink_ctx->cloud_conf = t_hive_sink.cloud_configuration;
-    sink_ctx->data_column_names = t_hive_sink.data_column_names;
-    sink_ctx->partition_column_names = t_hive_sink.partition_column_names;
-    sink_ctx->data_column_evaluators = ColumnExprEvaluator::from_exprs(data_exprs, runtime_state);
-    sink_ctx->partition_column_evaluators = ColumnExprEvaluator::from_exprs(partition_exprs, runtime_state);
-    auto* query_execution_services = runtime_state->query_execution_services();
-    sink_ctx->executor = query_execution_services->execution->pipeline_sink_io_pool;
-    sink_ctx->format = t_hive_sink.file_format;
-    sink_ctx->compression_type = t_hive_sink.compression_type;
-    if (t_hive_sink.__isset.target_max_file_size) {
-        sink_ctx->max_file_size = t_hive_sink.target_max_file_size;
-    }
-    if (t_hive_sink.__isset.text_file_desc) {
-        DCHECK(boost::iequals(t_hive_sink.file_format, formats::TEXTFILE));
-        sink_ctx->options[formats::CSVWriterOptions::COLUMN_TERMINATED_BY] = csv::DEFAULT_FIELD_DELIM;
-        sink_ctx->options[formats::CSVWriterOptions::LINE_TERMINATED_BY] = csv::DEFAULT_LINE_DELIM;
-        sink_ctx->options[formats::CSVWriterOptions::COLLECTION_DELIM] = csv::DEFAULT_COLLECTION_DELIM;
-        sink_ctx->options[formats::CSVWriterOptions::MAPKEY_DELIM] = csv::DEFAULT_MAPKEY_DELIM;
-        sink_ctx->options[formats::CSVWriterOptions::IS_HIVE] = "true";
-
-        // use customized value if specified
-        if (t_hive_sink.text_file_desc.__isset.field_delim) {
-            sink_ctx->options[formats::CSVWriterOptions::COLUMN_TERMINATED_BY] = t_hive_sink.text_file_desc.field_delim;
-        }
-        if (t_hive_sink.text_file_desc.__isset.line_delim) {
-            sink_ctx->options[formats::CSVWriterOptions::LINE_TERMINATED_BY] = t_hive_sink.text_file_desc.line_delim;
-        }
-        if (t_hive_sink.text_file_desc.__isset.collection_delim) {
-            sink_ctx->options[formats::CSVWriterOptions::COLLECTION_DELIM] =
-                    t_hive_sink.text_file_desc.collection_delim;
-        }
-        if (t_hive_sink.text_file_desc.__isset.mapkey_delim) {
-            sink_ctx->options[formats::CSVWriterOptions::MAPKEY_DELIM] = t_hive_sink.text_file_desc.mapkey_delim;
-        }
-    }
-    sink_ctx->runtime_state = runtime_state;
-
-    auto connector = connector::ConnectorRegistry::default_instance()->get(connector::Connector::HIVE);
-    ASSIGN_OR_RETURN(auto sink_provider,
-                     connector->create_sink_provider(starrocks::connector::ConnectorSinkProviderType::DATA, sink_ctx));
-    auto op = std::make_shared<pipeline::ConnectorSinkOperatorFactory>(context->next_operator_id(),
-                                                                       std::move(sink_provider), fragment_ctx);
-    size_t sink_dop = context->data_sink_dop();
-    if (t_hive_sink.partition_column_names.size() == 0 || t_hive_sink.is_static_partition_sink) {
-        auto ops = ::starrocks::pipeline::builder::maybe_interpolate_local_passthrough_exchange(
-                context, runtime_state, pipeline::Operator::s_pseudo_plan_node_id_for_final_sink, prev_operators,
-                sink_dop, pipeline::LocalExchanger::PassThroughType::SCALE);
-        ops.emplace_back(std::move(op));
-        context->add_pipeline(ops);
-    } else {
-        std::vector<ExprContext*> partition_expr_ctxs;
-        RETURN_IF_ERROR(ExprFactory::create_expr_trees(runtime_state->obj_pool(), partition_exprs, &partition_expr_ctxs,
-                                                       runtime_state));
-        auto ops = ::starrocks::pipeline::builder::interpolate_local_key_partition_exchange(
-                context, runtime_state, pipeline::Operator::s_pseudo_plan_node_id_for_final_sink, prev_operators,
-                partition_expr_ctxs, sink_dop);
-        ops.emplace_back(std::move(op));
-        context->add_pipeline(ops);
-    }
-
-    return Status::OK();
+    return HiveSinkBuilder(_t_output_expr).build(std::move(prev_operators), thrift_sink, context);
 }
 
 } // namespace starrocks
