+++ be/src/exec/data_sinks/hive_table_sink.h
@@ -14,12 +14,7 @@
 
 #pragma once
 
-#include "common/logging.h"
-#include "connector/hive/hive_chunk_sink.h"
-#include "exec/pipeline/sink/connector_sink_operator.h"
 #include "exec_primitive/data_sink.h"
-#include "formats/column_evaluator.h"
-#include "formats/csv/csv_file_writer.h"
 
 namespace starrocks {
