+++ be/src/exec/data_sinks/iceberg_sink_builder.cpp
@@ -1,0 +1,620 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "exec/data_sinks/iceberg_sink_builder.h"
+
+#include <unordered_map>
+
+#include "common/config_exec_fwd.h"
+#include "connector/iceberg/iceberg_connector.h"
+#include "connector/iceberg/iceberg_row_delta_sink.h"
+#include "connector/iceberg/iceberg_utils.h"
+#include "exec/pipeline/fragment_context.h"
+#include "exec/pipeline/pipeline_builder.h"
+#include "exec/pipeline/pipeline_builder_operators.h"
+#include "exec/pipeline/sink/connector_sink_operator.h"
+#include "exprs/expr.h"
+#include "exprs/expr_executor.h"
+#include "exprs/expr_factory.h"
+#include "formats/reserved_columns.h"
+#include "runtime/descriptor_helper.h"
+#include "runtime/descriptors_ext.h"
+#include "runtime/runtime_state.h"
+#include "runtime/service_contexts.h"
+#include "types/logical_type.h"
+
+namespace starrocks {
+
+namespace {
+
+std::unordered_map<std::string, formats::FileColumnId> build_top_level_field_id_map(
+        const std::vector<TIcebergSchemaField>& fields) {
+    std::unordered_map<std::string, formats::FileColumnId> field_ids_by_name;
+    auto field_ids = connector::IcebergUtils::generate_parquet_field_ids(fields);
+    field_ids_by_name.reserve(fields.size());
+    for (size_t i = 0; i < fields.size(); ++i) {
+        field_ids_by_name.emplace(fields[i].name, field_ids[i]);
+    }
+    return field_ids_by_name;
+}
+
+Status append_iceberg_sink_column(const SlotDescriptor* slot,
+                                  const std::unordered_map<std::string, formats::FileColumnId>& field_ids_by_name,
+                                  std::vector<std::string>* column_names,
+                                  std::vector<formats::FileColumnId>* parquet_field_ids) {
+    if (slot == nullptr) {
+        return Status::InternalError("Iceberg sink columns do not match output expressions");
+    }
+
+    const std::string col_name(slot->col_name());
+    column_names->emplace_back(col_name);
+
+    auto it = field_ids_by_name.find(col_name);
+    if (it != field_ids_by_name.end()) {
+        parquet_field_ids->push_back(it->second);
+        return Status::OK();
+    }
+
+    formats::FileColumnId field_id;
+    if (col_name == formats::kIcebergRowIdColumnName) {
+        field_id.field_id = formats::kIcebergRowIdColumnId;
+    } else if (col_name == formats::kIcebergLastUpdatedSequenceNumberColumnName) {
+        field_id.field_id = formats::kIcebergLastUpdatedSequenceNumberColumnId;
+    } else {
+        return Status::InternalError(
+                fmt::format("Iceberg sink field ids do not match output expressions, col_name={}", col_name));
+    }
+
+    parquet_field_ids->push_back(field_id);
+    return Status::OK();
+}
+
+} // namespace
+
+Status IcebergSinkBuilder::build(pipeline::OpFactories prev_operators, const TDataSink& thrift_sink,
+                                 pipeline::PipelineBuilderContext* context) const {
+    if (!thrift_sink.__isset.iceberg_table_sink) {
+        return Status::InternalError("Missing iceberg table sink");
+    }
+    auto* runtime_state = context->runtime_state();
+    auto* fragment_ctx = context->fragment_context();
+    TableDescriptor* table_desc =
+            runtime_state->desc_tbl().get_table_descriptor(thrift_sink.iceberg_table_sink.target_table_id);
+    if (table_desc == nullptr) {
+        return Status::InternalError("Missing Iceberg table descriptor");
+    }
+    auto* iceberg_table_desc = down_cast<IcebergTableDescriptor*>(table_desc);
+    auto& t_iceberg_sink = thrift_sink.iceberg_table_sink;
+
+    std::unique_ptr<connector::ConnectorSinkProvider> sink_provider;
+    std::vector<TExpr> partition_expr;
+    std::vector<std::string> transform_exprs = iceberg_table_desc->get_transform_exprs();
+
+    if (thrift_sink.type == TDataSinkType::ICEBERG_ROW_DELTA_SINK) {
+        RETURN_IF_ERROR(create_row_delta_sink_context(thrift_sink, runtime_state, context, iceberg_table_desc,
+                                                      sink_provider, partition_expr));
+    } else if (thrift_sink.type == TDataSinkType::ICEBERG_DELETE_SINK) {
+        RETURN_IF_ERROR(create_delete_sink_context(thrift_sink, runtime_state, context, iceberg_table_desc,
+                                                   sink_provider, partition_expr));
+    } else {
+        RETURN_IF_ERROR(create_data_sink_context(thrift_sink, runtime_state, context, iceberg_table_desc, sink_provider,
+                                                 partition_expr));
+    }
+
+    auto op = std::make_shared<pipeline::ConnectorSinkOperatorFactory>(context->next_operator_id(),
+                                                                       std::move(sink_provider), fragment_ctx);
+    size_t sink_dop = context->data_sink_dop();
+
+    if (iceberg_table_desc->is_unpartitioned_table() || t_iceberg_sink.is_static_partition_sink) {
+        auto ops = ::starrocks::pipeline::builder::maybe_interpolate_local_passthrough_exchange(
+                context, runtime_state, pipeline::Operator::s_pseudo_plan_node_id_for_final_sink, prev_operators,
+                sink_dop, pipeline::LocalExchanger::PassThroughType::SCALE);
+        ops.emplace_back(std::move(op));
+        context->add_pipeline(ops);
+    } else {
+        std::vector<ExprContext*> partition_expr_ctxs;
+
+        RETURN_IF_ERROR(ExprFactory::create_expr_trees(runtime_state->obj_pool(), partition_expr, &partition_expr_ctxs,
+                                                       runtime_state));
+
+        // Validate partition expressions were created successfully
+        for (int i = 0; i < partition_expr_ctxs.size(); i++) {
+            if (partition_expr_ctxs[i] == nullptr) {
+                return Status::InternalError(fmt::format("Partition expression at index {} is nullptr", i));
+            }
+            if (partition_expr_ctxs[i]->root() == nullptr) {
+                return Status::InternalError(fmt::format("Partition expression root at index {} is nullptr", i));
+            }
+        }
+
+        auto ops = ::starrocks::pipeline::builder::interpolate_local_key_partition_exchange(
+                context, runtime_state, pipeline::Operator::s_pseudo_plan_node_id_for_final_sink, prev_operators,
+                partition_expr_ctxs, sink_dop, transform_exprs);
+        ops.emplace_back(std::move(op));
+        context->add_pipeline(ops);
+    }
+
+    return Status::OK();
+}
+
+Status IcebergSinkBuilder::create_delete_sink_context(const TDataSink& thrift_sink, RuntimeState* runtime_state,
+                                                      pipeline::PipelineBuilderContext* /*context*/,
+                                                      IcebergTableDescriptor* iceberg_table_desc,
+                                                      std::unique_ptr<connector::ConnectorSinkProvider>& sink_provider,
+                                                      std::vector<TExpr>& partition_expr) const {
+    auto& t_iceberg_sink = thrift_sink.iceberg_table_sink;
+
+    // Create merge sink context for delete files
+    auto delete_sink_ctx = std::make_shared<connector::IcebergDeleteSinkContext>();
+    delete_sink_ctx->path = t_iceberg_sink.data_location;
+    delete_sink_ctx->cloud_configuration = t_iceberg_sink.cloud_configuration;
+    // Position-delete codec lives in delete_compression_type; fall back to
+    // compression_type so old FE planners (rolling upgrade with BE upgraded
+    // first) continue to pick up the right codec.
+    delete_sink_ctx->compression_type = t_iceberg_sink.__isset.delete_compression_type
+                                                ? t_iceberg_sink.delete_compression_type
+                                                : t_iceberg_sink.compression_type;
+    if (t_iceberg_sink.__isset.target_max_file_size) {
+        delete_sink_ctx->max_file_size = t_iceberg_sink.target_max_file_size;
+    }
+    delete_sink_ctx->tuple_desc_id = t_iceberg_sink.tuple_id;
+    auto* query_execution_services = runtime_state->query_execution_services();
+    delete_sink_ctx->executor = query_execution_services->execution->pipeline_sink_io_pool;
+    delete_sink_ctx->runtime_state = runtime_state;
+
+    // Delete files have columns: file_path and pos (row_position)
+    const auto& output_exprs = _output_exprs;
+    delete_sink_ctx->column_evaluators = ColumnExprEvaluator::from_exprs(output_exprs, runtime_state);
+    delete_sink_ctx->transform_exprs = iceberg_table_desc->get_transform_exprs();
+    delete_sink_ctx->output_exprs = output_exprs;
+
+    // Configure partition columns if table is partitioned
+    delete_sink_ctx->partition_column_names = iceberg_table_desc->partition_column_names();
+
+    // Build column name → slot reference map.
+    // Used by IcebergDeleteSink::add() to locate _file/_pos columns by slot_id,
+    // and by update_partition_expr_slot_refs_by_map() to bind partition expressions.
+    TupleDescriptor* tuple_desc = runtime_state->desc_tbl().get_tuple_descriptor(t_iceberg_sink.tuple_id);
+    if (tuple_desc == nullptr) {
+        return Status::InternalError(
+                fmt::format("Failed to find tuple descriptor with id {}", t_iceberg_sink.tuple_id));
+    }
+
+    const auto& slots = tuple_desc->slots();
+    if (slots.size() != output_exprs.size()) {
+        return Status::InternalError(fmt::format("Mismatched slot and output expression counts: {} vs {}", slots.size(),
+                                                 output_exprs.size()));
+    }
+
+    for (size_t i = 0; i < slots.size(); ++i) {
+        const auto* slot = slots[i];
+        const auto& expr = output_exprs[i];
+        for (const auto& node : expr.nodes) {
+            if (node.node_type == TExprNodeType::SLOT_REF && node.__isset.slot_ref) {
+                delete_sink_ctx->column_slot_map[std::string(slot->col_name())] = node;
+                break;
+            }
+        }
+    }
+
+    if (!iceberg_table_desc->is_unpartitioned_table()) {
+        partition_expr = iceberg_table_desc->get_partition_exprs();
+        const auto& partition_source_column_names = iceberg_table_desc->partition_source_column_names();
+
+        RETURN_IF_ERROR(update_partition_expr_slot_refs_by_map(partition_expr, delete_sink_ctx->column_slot_map,
+                                                               partition_source_column_names));
+        delete_sink_ctx->partition_evaluators = ColumnExprEvaluator::from_exprs(partition_expr, runtime_state);
+    }
+
+    connector::IcebergConnector iceberg_connector;
+    ASSIGN_OR_RETURN(auto provider, iceberg_connector.create_sink_provider(
+                                            starrocks::connector::ConnectorSinkProviderType::DELETE, delete_sink_ctx));
+    sink_provider = std::move(provider);
+
+    return Status::OK();
+}
+
+Status IcebergSinkBuilder::create_data_sink_context(const TDataSink& thrift_sink, RuntimeState* runtime_state,
+                                                    pipeline::PipelineBuilderContext* /*context*/,
+                                                    IcebergTableDescriptor* iceberg_table_desc,
+                                                    std::unique_ptr<connector::ConnectorSinkProvider>& sink_provider,
+                                                    std::vector<TExpr>& partition_expr) const {
+    auto& t_iceberg_sink = thrift_sink.iceberg_table_sink;
+
+    auto data_sink_ctx = std::make_shared<connector::IcebergChunkSinkContext>();
+    data_sink_ctx->path = t_iceberg_sink.__isset.data_location && !t_iceberg_sink.data_location.empty()
+                                  ? t_iceberg_sink.data_location
+                                  : t_iceberg_sink.location + connector::IcebergUtils::DATA_DIRECTORY;
+    data_sink_ctx->cloud_conf = t_iceberg_sink.cloud_configuration;
+    data_sink_ctx->partition_column_names = iceberg_table_desc->partition_column_names();
+    auto* query_execution_services = runtime_state->query_execution_services();
+    data_sink_ctx->executor = query_execution_services->execution->pipeline_sink_io_pool;
+    data_sink_ctx->format = t_iceberg_sink.file_format; // iceberg sink only supports parquet
+    data_sink_ctx->compression_type = t_iceberg_sink.compression_type;
+    if (t_iceberg_sink.__isset.target_max_file_size) {
+        data_sink_ctx->max_file_size = t_iceberg_sink.target_max_file_size;
+    }
+    data_sink_ctx->column_evaluators = ColumnExprEvaluator::from_exprs(_output_exprs, runtime_state);
+
+    size_t num_evaluators = data_sink_ctx->column_evaluators.size();
+    TupleDescriptor* tuple_desc = runtime_state->desc_tbl().get_tuple_descriptor(t_iceberg_sink.tuple_id);
+    const auto* slots = tuple_desc != nullptr ? &tuple_desc->slots() : nullptr;
+    if (slots == nullptr || slots->size() < num_evaluators) {
+        return Status::InternalError("Iceberg sink columns do not match output expressions");
+    }
+
+    // Build sink schema from the actual output tuple. TIcebergTable.columns may still contain
+    // hidden metadata columns such as _file/_pos that are not written by compaction.
+    auto field_ids_by_name = build_top_level_field_id_map(iceberg_table_desc->get_iceberg_schema()->fields);
+    data_sink_ctx->column_names.clear();
+    data_sink_ctx->parquet_field_ids.clear();
+    data_sink_ctx->column_names.reserve(num_evaluators);
+    data_sink_ctx->parquet_field_ids.reserve(num_evaluators);
+    for (size_t i = 0; i < num_evaluators; ++i) {
+        RETURN_IF_ERROR(append_iceberg_sink_column((*slots)[i], field_ids_by_name, &data_sink_ctx->column_names,
+                                                   &data_sink_ctx->parquet_field_ids));
+    }
+
+    if (data_sink_ctx->column_names.size() != num_evaluators ||
+        data_sink_ctx->parquet_field_ids.size() != num_evaluators) {
+        return Status::InternalError("Iceberg sink schema metadata does not match output expressions");
+    }
+
+    data_sink_ctx->transform_exprs = iceberg_table_desc->get_transform_exprs();
+    data_sink_ctx->runtime_state = runtime_state;
+    data_sink_ctx->tuple_desc_id = t_iceberg_sink.tuple_id;
+    auto& sort_order = iceberg_table_desc->sort_order();
+    if (!sort_order.sort_key_idxes.empty()) {
+        data_sink_ctx->sort_ordering = std::make_shared<connector::SortOrdering>();
+        data_sink_ctx->sort_ordering->sort_key_idxes.assign(sort_order.sort_key_idxes.begin(),
+                                                            sort_order.sort_key_idxes.end());
+        data_sink_ctx->sort_ordering->sort_descs.descs.reserve(sort_order.sort_key_idxes.size());
+        for (size_t idx = 0; idx < sort_order.sort_key_idxes.size(); ++idx) {
+            bool is_asc = idx < sort_order.is_ascs.size() ? sort_order.is_ascs[idx] : true;
+            bool is_null_first = false;
+            if (idx < sort_order.is_null_firsts.size()) {
+                is_null_first = sort_order.is_null_firsts[idx];
+            } else if (is_asc) {
+                // If ascending, nulls are first by default
+                // If descending, nulls are last by default
+                is_null_first = true;
+            }
+            data_sink_ctx->sort_ordering->sort_descs.descs.emplace_back(is_asc, is_null_first);
+        }
+    }
+
+    connector::IcebergConnector iceberg_connector;
+    ASSIGN_OR_RETURN(auto provider, iceberg_connector.create_sink_provider(
+                                            starrocks::connector::ConnectorSinkProviderType::DATA, data_sink_ctx));
+    sink_provider = std::move(provider);
+
+    if (iceberg_table_desc->is_unpartitioned_table()) {
+        //do nothing
+    } else if (t_iceberg_sink.is_static_partition_sink) {
+        for (const auto& index : iceberg_table_desc->partition_source_index_in_schema()) {
+            if (index < 0 || index >= _output_exprs.size()) {
+                return Status::InternalError(fmt::format("Invalid partition index: {}", index));
+            }
+            partition_expr.push_back(_output_exprs[index]);
+        }
+        data_sink_ctx->partition_evaluators = ColumnExprEvaluator::from_exprs(partition_expr, runtime_state);
+    } else {
+        auto source_column_index = iceberg_table_desc->partition_source_index_in_schema();
+        partition_expr = iceberg_table_desc->get_partition_exprs();
+
+        //for 3.5 fe -> 4.0 be compact, try to set this.
+        if (partition_expr.empty()) {
+            auto output_expr = _output_exprs;
+            for (const auto& index : source_column_index) {
+                if (index < 0 || index >= _output_exprs.size()) {
+                    return Status::InternalError(fmt::format("Invalid partition index: {}", index));
+                }
+                partition_expr.push_back(output_expr[index]);
+            }
+        }
+
+        int idx = 0;
+        for (auto& part_expr : partition_expr) {
+            int index = source_column_index[idx];
+            //check index is valid for output_expr
+            if (index < 0 || index >= _output_exprs.size()) {
+                return Status::InternalError(fmt::format("Invalid partition index: {}", index));
+            }
+            auto slot_ref = _output_exprs[index];
+            for (auto& node : part_expr.nodes) {
+                if (node.node_type == TExprNodeType::SLOT_REF) {
+                    node = slot_ref.nodes[0];
+                    break;
+                }
+            }
+            idx++;
+        }
+        data_sink_ctx->partition_evaluators = ColumnExprEvaluator::from_exprs(partition_expr, runtime_state);
+    }
+
+    return Status::OK();
+}
+
+Status IcebergSinkBuilder::create_row_delta_sink_context(
+        const TDataSink& thrift_sink, RuntimeState* runtime_state, pipeline::PipelineBuilderContext* /*context*/,
+        IcebergTableDescriptor* iceberg_table_desc, std::unique_ptr<connector::ConnectorSinkProvider>& sink_provider,
+        std::vector<TExpr>& partition_expr) const {
+    auto& t_iceberg_sink = thrift_sink.iceberg_table_sink;
+
+    // Resolve tuple descriptor
+    TupleDescriptor* tuple_desc = runtime_state->desc_tbl().get_tuple_descriptor(t_iceberg_sink.tuple_id);
+    if (tuple_desc == nullptr) {
+        return Status::InternalError(
+                fmt::format("Failed to find tuple descriptor with id {}", t_iceberg_sink.tuple_id));
+    }
+
+    const auto& slots = tuple_desc->slots();
+    const auto& output_exprs = _output_exprs;
+    if (slots.size() != output_exprs.size()) {
+        return Status::InternalError(fmt::format("Mismatched slot and output expression counts: {} vs {}", slots.size(),
+                                                 output_exprs.size()));
+    }
+
+    // Determine column layout indices:
+    //   ROW_DELTA_UPDATE: [_file, _pos, data_col1, ..., data_colN]
+    //   ROW_DELTA_MIXED:  [_file, _pos, data_col1, ..., data_colN, op_code]
+    // No partition prefix — partition columns are part of data_cols (full table schema).
+    const size_t total_slots = slots.size();
+    const int32_t data_column_start = 2; // after _file and _pos
+    if (!t_iceberg_sink.__isset.write_mode) {
+        return Status::InternalError("Iceberg row delta sink requires write_mode");
+    }
+    const auto write_mode = t_iceberg_sink.write_mode;
+    if (write_mode != TIcebergWriteMode::ROW_DELTA_UPDATE && write_mode != TIcebergWriteMode::ROW_DELTA_MIXED) {
+        return Status::InternalError(
+                fmt::format("Invalid Iceberg row delta write_mode: {}", static_cast<int>(write_mode)));
+    }
+    const bool has_routing_column = write_mode == TIcebergWriteMode::ROW_DELTA_MIXED;
+    const int32_t op_code_index = has_routing_column ? static_cast<int32_t>(total_slots) - 1 : -1;
+    const int32_t data_column_end =
+            has_routing_column ? static_cast<int32_t>(total_slots) - 1 : static_cast<int32_t>(total_slots);
+
+    if (data_column_end <= data_column_start) {
+        return Status::InternalError(
+                fmt::format("Iceberg row delta layout has no data columns: total_slots={}, data_column_start={}, "
+                            "write_mode={}",
+                            total_slots, data_column_start, static_cast<int>(write_mode)));
+    }
+
+    if (has_routing_column) {
+        if (slots[op_code_index]->type().type != TYPE_TINYINT) {
+            return Status::InternalError(fmt::format("Iceberg row delta op column must be TINYINT, actual={}",
+                                                     type_to_string(slots[op_code_index]->type().type)));
+        }
+    }
+
+    // --- Create delete sub-context ---
+    auto delete_sink_ctx = std::make_shared<connector::IcebergDeleteSinkContext>();
+    delete_sink_ctx->writer_tag = "delete";
+    delete_sink_ctx->path = t_iceberg_sink.data_location;
+    delete_sink_ctx->cloud_configuration = t_iceberg_sink.cloud_configuration;
+    // Row-delta sinks carry distinct codecs for data vs delete files; fall back to
+    // the data codec when the planner did not populate delete_compression_type.
+    delete_sink_ctx->compression_type = t_iceberg_sink.__isset.delete_compression_type
+                                                ? t_iceberg_sink.delete_compression_type
+                                                : t_iceberg_sink.compression_type;
+    if (t_iceberg_sink.__isset.target_max_file_size) {
+        delete_sink_ctx->max_file_size = t_iceberg_sink.target_max_file_size;
+    }
+    delete_sink_ctx->tuple_desc_id = t_iceberg_sink.tuple_id;
+    auto* query_execution_services = runtime_state->query_execution_services();
+    delete_sink_ctx->executor = query_execution_services->execution->pipeline_sink_io_pool;
+    delete_sink_ctx->runtime_state = runtime_state;
+
+    // Pass the full output_exprs list (one per tuple slot) so
+    // IcebergDeleteSinkProvider::create_sink()'s strict
+    // `tuple_desc->slots().size() == ctx->output_exprs.size()` DCHECK holds in
+    // debug builds. The delete sub-sink only consumes evaluators [0]=_file and
+    // [1]=_pos at runtime; trailing data-column / op_code evaluators are held
+    // but not used.
+    delete_sink_ctx->column_evaluators = ColumnExprEvaluator::from_exprs(output_exprs, runtime_state);
+    delete_sink_ctx->transform_exprs = iceberg_table_desc->get_transform_exprs();
+    delete_sink_ctx->output_exprs = output_exprs;
+
+    // Configure partition columns for delete context
+    delete_sink_ctx->partition_column_names = iceberg_table_desc->partition_column_names();
+
+    // Build column name → slot reference map from all columns (except mixed-mode op_code).
+    // IcebergDeleteSink::add() uses this to locate _file/_pos by slot_id,
+    // and update_partition_expr_slot_refs_by_map() uses it to bind partition exprs.
+    // Since we pass the full original chunk to sub-sinks (not sub-chunks),
+    // all slot_ids are available at runtime.
+    for (int32_t i = 0; i < static_cast<int32_t>(total_slots); ++i) {
+        if (i == op_code_index) {
+            continue;
+        }
+        const auto* slot = slots[i];
+        const auto& expr = output_exprs[i];
+        for (const auto& node : expr.nodes) {
+            if (node.node_type == TExprNodeType::SLOT_REF && node.__isset.slot_ref) {
+                delete_sink_ctx->column_slot_map[std::string(slot->col_name())] = node;
+                break;
+            }
+        }
+    }
+
+    // --- Create data sub-context ---
+    auto data_sink_ctx = std::make_shared<connector::IcebergChunkSinkContext>();
+    data_sink_ctx->writer_tag = "data";
+    data_sink_ctx->path = t_iceberg_sink.__isset.data_location && !t_iceberg_sink.data_location.empty()
+                                  ? t_iceberg_sink.data_location
+                                  : t_iceberg_sink.location + connector::IcebergUtils::DATA_DIRECTORY;
+    data_sink_ctx->cloud_conf = t_iceberg_sink.cloud_configuration;
+    data_sink_ctx->partition_column_names = iceberg_table_desc->partition_column_names();
+    data_sink_ctx->executor = query_execution_services->execution->pipeline_sink_io_pool;
+    data_sink_ctx->format = t_iceberg_sink.file_format;
+    data_sink_ctx->compression_type = t_iceberg_sink.compression_type;
+    if (t_iceberg_sink.__isset.target_max_file_size) {
+        data_sink_ctx->max_file_size = t_iceberg_sink.target_max_file_size;
+    }
+
+    // Data output_exprs: data columns only; mixed-mode op_code is private to the
+    // composite row-delta sink and must not be written into data files.
+    std::vector<TExpr> data_output_exprs(output_exprs.begin() + data_column_start,
+                                         output_exprs.begin() + data_column_end);
+    data_sink_ctx->column_evaluators = ColumnExprEvaluator::from_exprs(data_output_exprs, runtime_state);
+
+    size_t num_data_evaluators = data_sink_ctx->column_evaluators.size();
+
+    // Build sink schema from data columns
+    auto field_ids_by_name = build_top_level_field_id_map(iceberg_table_desc->get_iceberg_schema()->fields);
+    data_sink_ctx->column_names.clear();
+    data_sink_ctx->parquet_field_ids.clear();
+    data_sink_ctx->column_names.reserve(num_data_evaluators);
+    data_sink_ctx->parquet_field_ids.reserve(num_data_evaluators);
+    for (size_t i = 0; i < num_data_evaluators; ++i) {
+        RETURN_IF_ERROR(append_iceberg_sink_column(slots[data_column_start + i], field_ids_by_name,
+                                                   &data_sink_ctx->column_names, &data_sink_ctx->parquet_field_ids));
+    }
+
+    if (data_sink_ctx->column_names.size() != num_data_evaluators ||
+        data_sink_ctx->parquet_field_ids.size() != num_data_evaluators) {
+        return Status::InternalError("Iceberg row delta sink schema metadata does not match output expressions");
+    }
+
+    data_sink_ctx->transform_exprs = iceberg_table_desc->get_transform_exprs();
+    data_sink_ctx->runtime_state = runtime_state;
+
+    // Create a data-only tuple descriptor for the data sub-sink.
+    // The full row-delta tuple includes _file, _pos, data_cols and optionally op_code,
+    // but the data writer only has evaluators for data_cols. SpillPartitionChunkWriter
+    // sizes merged chunks from tuple_desc->slots().size(), so the full tuple causes
+    // out-of-bounds.
+    {
+        TSlotDescriptorBuilder slot_builder;
+        TTupleDescriptorBuilder tuple_builder;
+        for (size_t i = 0; i < num_data_evaluators; ++i) {
+            const auto* slot = slots[data_column_start + i];
+            // Pass the full TypeDescriptor (not just slot->type().type) so complex types
+            // (ARRAY/MAP/STRUCT) keep their children and scalar types keep len/precision/scale.
+            // A childless ARRAY/MAP descriptor crashes TypeDescriptor::to_thrift() on children[0].
+            tuple_builder.add_slot(slot_builder.type(slot->type())
+                                           .nullable(slot->is_nullable())
+                                           .is_materialized(true)
+                                           .column_name(std::string(slot->col_name()))
+                                           .build());
+        }
+        TDescriptorTableBuilder desc_tbl_builder;
+        tuple_builder.build(&desc_tbl_builder);
+        TDescriptorTable t_desc_tbl = desc_tbl_builder.desc_tbl();
+        DescriptorTbl* data_desc_tbl = nullptr;
+        RETURN_IF_ERROR(DescriptorTbl::create(runtime_state, runtime_state->obj_pool(), t_desc_tbl, &data_desc_tbl,
+                                              config::vector_chunk_size));
+        data_sink_ctx->override_tuple_desc = data_desc_tbl->get_tuple_descriptor(0);
+        DCHECK(data_sink_ctx->override_tuple_desc != nullptr);
+    }
+    data_sink_ctx->tuple_desc_id = t_iceberg_sink.tuple_id;
+
+    // Sort ordering for data files
+    auto& sort_order = iceberg_table_desc->sort_order();
+    if (!sort_order.sort_key_idxes.empty()) {
+        data_sink_ctx->sort_ordering = std::make_shared<connector::SortOrdering>();
+        data_sink_ctx->sort_ordering->sort_key_idxes.assign(sort_order.sort_key_idxes.begin(),
+                                                            sort_order.sort_key_idxes.end());
+        data_sink_ctx->sort_ordering->sort_descs.descs.reserve(sort_order.sort_key_idxes.size());
+        for (size_t idx = 0; idx < sort_order.sort_key_idxes.size(); ++idx) {
+            bool is_asc = idx < sort_order.is_ascs.size() ? sort_order.is_ascs[idx] : true;
+            bool is_null_first = false;
+            if (idx < sort_order.is_null_firsts.size()) {
+                is_null_first = sort_order.is_null_firsts[idx];
+            } else if (is_asc) {
+                is_null_first = true;
+            }
+            data_sink_ctx->sort_ordering->sort_descs.descs.emplace_back(is_asc, is_null_first);
+        }
+    }
+
+    // --- Set up partition expressions for both contexts ---
+    if (!iceberg_table_desc->is_unpartitioned_table()) {
+        partition_expr = iceberg_table_desc->get_partition_exprs();
+        const auto& partition_source_column_names = iceberg_table_desc->partition_source_column_names();
+
+        // Update partition expressions using column slot map (same approach as delete sink)
+        RETURN_IF_ERROR(update_partition_expr_slot_refs_by_map(partition_expr, delete_sink_ctx->column_slot_map,
+                                                               partition_source_column_names));
+        delete_sink_ctx->partition_evaluators = ColumnExprEvaluator::from_exprs(partition_expr, runtime_state);
+        data_sink_ctx->partition_evaluators = ColumnExprEvaluator::from_exprs(partition_expr, runtime_state);
+    }
+
+    // --- Assemble IcebergRowDeltaSinkContext ---
+    auto row_delta_ctx = std::make_shared<connector::IcebergRowDeltaSinkContext>();
+    row_delta_ctx->delete_sink_ctx = delete_sink_ctx;
+    row_delta_ctx->data_sink_ctx = data_sink_ctx;
+    row_delta_ctx->op_code_index = op_code_index;
+
+    connector::IcebergConnector iceberg_connector;
+    ASSIGN_OR_RETURN(auto provider, iceberg_connector.create_sink_provider(
+                                            starrocks::connector::ConnectorSinkProviderType::ROW_DELTA, row_delta_ctx));
+    sink_provider = std::move(provider);
+
+    return Status::OK();
+}
+
+// Updates partition expression slot references using the column slot map.
+// For each partition expression, replaces the slot reference node with the correct
+// slot reference from the column_slot_map based on the partition source column name.
+//
+// Parameters:
+//   partition_expr - The partition expressions to update (modified in-place)
+//   column_slot_map - Mapping from column name to slot reference node
+//   partition_source_column_names - Names of source columns for partitioning
+//
+// Returns Status::OK() on success, or an error if a required slot reference is missing.
+Status IcebergSinkBuilder::update_partition_expr_slot_refs_by_map(
+        std::vector<TExpr>& partition_expr, const std::unordered_map<std::string, TExprNode>& column_slot_map,
+        const std::vector<std::string>& partition_source_column_names) const {
+    // Validate input sizes match
+    if (partition_expr.size() != partition_source_column_names.size()) {
+        return Status::InternalError(fmt::format("Mismatched partition expression and column name counts: {} vs {}",
+                                                 partition_expr.size(), partition_source_column_names.size()));
+    }
+
+    // Update each partition expression's slot reference
+    for (size_t i = 0; i < partition_expr.size(); ++i) {
+        const std::string& source_col_name = partition_source_column_names[i];
+
+        // Find the slot reference for this source column
+        auto it = column_slot_map.find(source_col_name);
+        if (it == column_slot_map.end()) {
+            return Status::InternalError(
+                    fmt::format("Could not find slot reference for partition column: {}", source_col_name));
+        }
+
+        const TExprNode& slot_info = it->second;
+        bool slot_ref_updated = false;
+
+        // Update the slot reference in the partition expression
+        for (auto& node : partition_expr[i].nodes) {
+            if (node.node_type == TExprNodeType::SLOT_REF && node.__isset.slot_ref) {
+                node = slot_info;
+                slot_ref_updated = true;
+                break;
+            }
+        }
+
+        if (!slot_ref_updated) {
+            return Status::InternalError(
+                    fmt::format("No slot reference found in partition expression for column: {}", source_col_name));
+        }
+    }
+
+    return Status::OK();
+}
+
+} // namespace starrocks
