+++ be/src/exec/data_sinks/table_function_sink_builder.cpp
@@ -1,0 +1,148 @@
+// Copyright 2021-present StarRocks, Inc. All rights reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "exec/data_sinks/table_function_sink_builder.h"
+
+#include "connector/connector_registry.h"
+#include "connector/file/file_chunk_sink.h"
+#include "exec/pipeline/fragment_context.h"
+#include "exec/pipeline/pipeline_builder.h"
+#include "exec/pipeline/pipeline_builder_operators.h"
+#include "exec/pipeline/sink/connector_sink_operator.h"
+#include "exprs/expr_factory.h"
+#include "fmt/format.h"
+#include "formats/column_evaluator.h"
+#include "formats/csv/csv_file_writer.h"
+#include "formats/parquet/parquet_file_writer.h"
+#include "glog/logging.h"
+#include "runtime/runtime_state.h"
+#include "runtime/service_contexts.h"
+
+namespace starrocks {
+
+Status TableFunctionSinkBuilder::build(pipeline::OpFactories prev_operators, const TDataSink& thrift_sink,
+                                       pipeline::PipelineBuilderContext* context) const {
+    if (!thrift_sink.__isset.table_function_table_sink) {
+        return Status::InternalError("Missing table function table sink");
+    }
+    auto runtime_state = context->runtime_state();
+    auto fragment_ctx = context->fragment_context();
+    const auto& output_exprs = _output_exprs;
+
+    if (!thrift_sink.table_function_table_sink.__isset.target_table) {
+        return Status::InternalError("Missing table function target table");
+    }
+    if (!thrift_sink.table_function_table_sink.__isset.cloud_configuration) {
+        return Status::InternalError("Missing table function cloud configuration");
+    }
+
+    const auto& target_table = thrift_sink.table_function_table_sink.target_table;
+    if (!target_table.__isset.path || !target_table.__isset.file_format || !target_table.__isset.columns ||
+        !target_table.__isset.write_single_file) {
+        return Status::InternalError("Incomplete table function target table");
+    }
+    if (target_table.columns.size() != output_exprs.size()) {
+        return Status::InternalError("Table function columns do not match output expressions");
+    }
+
+    std::vector<std::string> column_names;
+    column_names.reserve(target_table.columns.size());
+    for (const auto& column : target_table.columns) {
+        column_names.push_back(column.column_name);
+    }
+
+    // prepare sink context
+    auto sink_ctx = std::make_shared<connector::FileChunkSinkContext>();
+    sink_ctx->path = target_table.path;
+    sink_ctx->cloud_conf = thrift_sink.table_function_table_sink.cloud_configuration;
+    sink_ctx->column_names = std::move(column_names);
+    if (target_table.__isset.partition_column_ids) {
+        sink_ctx->partition_column_indices = target_table.partition_column_ids;
+        for (auto id : sink_ctx->partition_column_indices) {
+            if (id < 0 || static_cast<size_t>(id) >= output_exprs.size()) {
+                return Status::InternalError(fmt::format("Invalid table function partition column index: {}", id));
+            }
+        }
+    }
+    auto* query_execution_services = runtime_state->query_execution_services();
+    sink_ctx->executor = query_execution_services->execution->pipeline_sink_io_pool;
+    sink_ctx->format = target_table.file_format;
+    if (target_table.__isset.target_max_file_size) {
+        sink_ctx->max_file_size = target_table.target_max_file_size;
+    }
+    if (target_table.__isset.write_single_file && target_table.write_single_file) {
+        sink_ctx->max_file_size = INT64_MAX;
+    }
+    sink_ctx->compression_type = target_table.compression_type;
+    sink_ctx->column_evaluators = ColumnExprEvaluator::from_exprs(output_exprs, runtime_state);
+    sink_ctx->runtime_state = runtime_state;
+    if (target_table.__isset.csv_column_seperator) {
+        sink_ctx->options[formats::CSVWriterOptions::COLUMN_TERMINATED_BY] = target_table.csv_column_seperator;
+    }
+    if (target_table.__isset.csv_row_delimiter) {
+        sink_ctx->options[formats::CSVWriterOptions::LINE_TERMINATED_BY] = target_table.csv_row_delimiter;
+    }
+    if (target_table.__isset.csv_include_header && target_table.csv_include_header) {
+        sink_ctx->options[formats::CSVWriterOptions::INCLUDE_HEADER] = "true";
+    }
+    if (target_table.__isset.csv_enclose) {
+        sink_ctx->options[formats::CSVWriterOptions::ENCLOSE] =
+                std::string(1, static_cast<char>(target_table.csv_enclose));
+    }
+    if (target_table.__isset.csv_escape) {
+        sink_ctx->options[formats::CSVWriterOptions::ESCAPE] =
+                std::string(1, static_cast<char>(target_table.csv_escape));
+    }
+    if (target_table.__isset.parquet_use_legacy_encoding && target_table.parquet_use_legacy_encoding) {
+        sink_ctx->options[formats::ParquetWriterOptions::USE_LEGACY_DECIMAL_ENCODING] = "true";
+        sink_ctx->options[formats::ParquetWriterOptions::USE_INT96_TIMESTAMP_ENCODING] = "true";
+    }
+    if (target_table.__isset.parquet_options && target_table.parquet_options.__isset.version) {
+        sink_ctx->options[formats::ParquetWriterOptions::VERSION] = target_table.parquet_options.version;
+    }
+
+    auto connector = connector::ConnectorRegistry::default_instance()->get(connector::Connector::FILE);
+    ASSIGN_OR_RETURN(auto sink_provider,
+                     connector->create_sink_provider(starrocks::connector::ConnectorSinkProviderType::DATA, sink_ctx));
+    auto op = std::make_shared<pipeline::ConnectorSinkOperatorFactory>(context->next_operator_id(),
+                                                                       std::move(sink_provider), fragment_ctx);
+
+    size_t sink_dop = target_table.write_single_file ? 1 : context->data_sink_dop();
+    if (sink_ctx->partition_column_indices.empty()) {
+        auto ops = ::starrocks::pipeline::builder::maybe_interpolate_local_passthrough_exchange(
+                context, runtime_state, pipeline::Operator::s_pseudo_plan_node_id_for_final_sink, prev_operators,
+                sink_dop, pipeline::LocalExchanger::PassThroughType::SCALE);
+        ops.emplace_back(std::move(op));
+        context->add_pipeline(ops);
+
+    } else {
+        std::vector<TExpr> partition_exprs;
+        partition_exprs.reserve(target_table.partition_column_ids.size());
+        for (auto id : target_table.partition_column_ids) {
+            partition_exprs.push_back(output_exprs[id]);
+        }
+        std::vector<ExprContext*> partition_expr_ctxs;
+        RETURN_IF_ERROR(ExprFactory::create_expr_trees(runtime_state->obj_pool(), partition_exprs, &partition_expr_ctxs,
+                                                       runtime_state));
+        auto ops = ::starrocks::pipeline::builder::interpolate_local_key_partition_exchange(
+                context, runtime_state, pipeline::Operator::s_pseudo_plan_node_id_for_final_sink, prev_operators,
+                partition_expr_ctxs, sink_dop);
+        ops.emplace_back(std::move(op));
+        context->add_pipeline(ops);
+    }
+
+    return Status::OK();
+}
+
+} // namespace starrocks
